refactor: seperate ocr and api module
