+++ libro_server/libro_ai_server/src/book_recognition/book_recognition/modules/ocr_module.py
@@ -1,0 +1,49 @@
+from book_recognition.config.config import pkg_path
+from paddleocr import PaddleOCR
+
+from PIL import Image, ImageDraw, ImageFont
+import cv2
+import os
+import numpy as np
+from pathlib import Path
+
+
+# ex) image_list = [image1, image2, ...]와 같은 형태. 각 이미지는 cv 형식
+# 책들의 ocr 리스트를 생성
+def run_ocr(image_list):
+    # PaddleOCR 리더 생성(한글 'korean' 설정)
+    ocr = PaddleOCR(lang='korean')
+    search_keyword_list = []
+
+    for img in image_list:
+        img = cv2.resize(img, None, fx=2, fy=2, interpolation=cv2.INTER_CUBIC)
+
+        # PIL 형식으로 변환
+        img = Image.fromarray(cv2.cvtColor(img, cv2.COLOR_BGR2RGB))
+        
+        draw = ImageDraw.Draw(img)
+        font = ImageFont.truetype(os.path.join(pkg_path, 'resources', 'MaruBuri-Bold.ttf'), 20)
+
+        # ocr로 추출한 단어 저장
+        extracted_words = ""
+
+        # 텍스트 검출 및 박스 그리기
+        for line in ocr.ocr(np.array(img), cls=True):
+            for word_info in line:
+                box, (text, score) = word_info[0], word_info[1]
+                top_left, bottom_right = tuple(map(int, box[0])), tuple(map(int, box[2]))
+
+                extracted_words += text
+
+                # 박스 그리기
+                draw.rectangle([top_left, bottom_right], outline=(250, 0, 0), width=2)
+                draw.text(top_left, text, font=font, fill=(0, 0, 255))
+                
+        search_keyword_list.append(extracted_words)
+        
+        # ocr 결과 이미지 저장
+        save_path = Path(f'./ocr_test/{extracted_words}.jpg')
+        save_path.parent.mkdir(parents=True, exist_ok=True)
+        img.save(save_path)
+
+    return search_keyword_list
